Revert "Revert "PlatformMerger""
This reverts commit a4741b5bf81def42b7099ee307041c556178b5ad.
