+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/handlers/v2/DescriptorHandler.java
@@ -62,7 +62,7 @@ public class DescriptorHandler {
         return new ModulesSorter(descriptor, this, dependencyTypeProperty, hardDependencyType);
     }
     
-    public ResourceType findPlatformResourceType(Platform platform, String resourceTypeName) {
+    public ResourceType findResourceType(Platform platform, String resourceTypeName) {
         for (ResourceType resourceType : platform.getResourceTypes2()) {
             if (resourceType.getName()
                 .equals(resourceTypeName)) {
@@ -72,7 +72,7 @@ public class DescriptorHandler {
         return null;
     }
 
-    public ModuleType findPlatformModuleType(Platform platform, String moduleTypeName) {
+    public ModuleType findModuleType(Platform platform, String moduleTypeName) {
         for (ModuleType moduleType : platform.getModuleTypes2()) {
             if (moduleType.getName()
                 .equals(moduleTypeName)) {
