+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/mergers/v2/PlatformMerger.java
@@ -3,32 +3,32 @@ package com.sap.cloud.lm.sl.mta.mergers.v2;
 import com.sap.cloud.lm.sl.common.util.MapUtil;
 import com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorHandler;
 import com.sap.cloud.lm.sl.mta.model.ElementContext;
+import com.sap.cloud.lm.sl.mta.model.Visitor;
+import com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor;
+import com.sap.cloud.lm.sl.mta.model.v2.Module;
 import com.sap.cloud.lm.sl.mta.model.v2.Platform;
 import com.sap.cloud.lm.sl.mta.model.v2.ModuleType;
 import com.sap.cloud.lm.sl.mta.model.v2.ResourceType;
+import com.sap.cloud.lm.sl.mta.model.v2.Resource;
 
-public class PlatformMerger extends com.sap.cloud.lm.sl.mta.mergers.v1.PlatformMerger {
+public class PlatformMerger extends Visitor {
+
+    protected final Platform platform;
+    protected final DescriptorHandler handler;
 
     public PlatformMerger(Platform platform, DescriptorHandler handler) {
-        super(platform, handler);
+        this.platform = platform;
+        this.handler = handler;
     }
 
     @Override
-    public void visit(ElementContext context, com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor descriptor) {
-        visit(context, (com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor) descriptor);
-    }
-
-    public void visit(ElementContext context, com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor descriptor) {
-        com.sap.cloud.lm.sl.mta.model.v2.Platform v2Platform = (Platform) platform;
+    public void visit(ElementContext context, DeploymentDescriptor descriptor) {
+        Platform v2Platform = platform;
         descriptor.setParameters(MapUtil.merge(v2Platform.getParameters(), descriptor.getParameters()));
     }
 
     @Override
-    public void visit(ElementContext context, com.sap.cloud.lm.sl.mta.model.v1.Resource resource) {
-        visit(context, (com.sap.cloud.lm.sl.mta.model.v2.Resource) resource);
-    }
-
-    public void visit(ElementContext context, com.sap.cloud.lm.sl.mta.model.v2.Resource resource) {
+    public void visit(ElementContext context, Resource resource) {
         if (resource.getType() == null) {
             return;
         }
@@ -39,10 +39,6 @@ public class PlatformMerger extends com.sap.cloud.lm.sl.mta.mergers.v1.PlatformM
         resource.setParameters(MapUtil.merge(resourceType.getParameters(), resource.getParameters()));
     }
 
-    @Override
-    public void visit(ElementContext context, com.sap.cloud.lm.sl.mta.model.v1.Module module) {
-        visit(context, (com.sap.cloud.lm.sl.mta.model.v2.Module) module);
-    }
 
     public void visit(ElementContext context, com.sap.cloud.lm.sl.mta.model.v2.Module module) {
         ModuleType moduleType = (ModuleType) handler.findModuleType(platform, module.getType());
